CODE_PROPRE_PROF_READY/requirements-dev.txt
customtkinter==5.2.2
pygame-ce==2.5.7
mariadb
pyinstaller
Pillow
